Broke down into a proper app
